+++ lib/src/v2/document.dart
@@ -1,11 +1,11 @@
-import 'package:codable/cast.dart' as cast;
-import 'package:open_api/src/object.dart';
-import 'package:open_api/src/v2/metadata.dart';
-import 'package:open_api/src/v2/parameter.dart';
-import 'package:open_api/src/v2/path.dart';
-import 'package:open_api/src/v2/response.dart';
-import 'package:open_api/src/v2/schema.dart';
-import 'package:open_api/src/v2/security.dart';
+import 'package:codable_forked/cast.dart' as cast;
+import 'package:open_api_forked/src/object.dart';
+import 'package:open_api_forked/src/v2/metadata.dart';
+import 'package:open_api_forked/src/v2/parameter.dart';
+import 'package:open_api_forked/src/v2/path.dart';
+import 'package:open_api_forked/src/v2/response.dart';
+import 'package:open_api_forked/src/v2/schema.dart';
+import 'package:open_api_forked/src/v2/security.dart';
 
 /// Represents an OpenAPI 2.0 specification.
 class APIDocument extends APIObject {
@@ -18,21 +18,21 @@ class APIDocument extends APIObject {
   }
 
   String version = "2.0";
-  APIInfo info = new APIInfo();
-  String host;
-  String basePath;
+  APIInfo? info = APIInfo();
+  String? host;
+  String? basePath;
 
-  List<APITag> tags = [];
-  List<String> schemes = [];
-  List<String> consumes = [];
-  List<String> produces = [];
-  List<Map<String, List<String>>> security = [];
+  List<APITag?>? tags = [];
+  List<String>? schemes = [];
+  List<String>? consumes = [];
+  List<String>? produces = [];
+  List<Map<String, List<String?>>?> security = [];
 
-  Map<String, APIPath> paths = {};
-  Map<String, APIResponse> responses = {};
-  Map<String, APIParameter> parameters = {};
-  Map<String, APISchemaObject> definitions = {};
-  Map<String, APISecurityScheme> securityDefinitions = {};
+  Map<String, APIPath?>? paths = {};
+  Map<String, APIResponse?>? responses = {};
+  Map<String, APIParameter?>? parameters = {};
+  Map<String, APISchemaObject?>? definitions = {};
+  Map<String, APISecurityScheme?>? securityDefinitions = {};
 
   Map<String, dynamic> asMap() {
     return KeyedArchive.archive(this, allowReferences: true);
@@ -57,13 +57,15 @@ class APIDocument extends APIObject {
     produces = object["produces"];
     security = object["security"];
 
-    info = object.decodeObject("info", () => new APIInfo());
-    tags = object.decodeObjects("tags", () => new APITag());
-    paths = object.decodeObjectMap("paths", () => new APIPath());
-    responses = object.decodeObjectMap("responses", () => new APIResponse());
-    parameters = object.decodeObjectMap("parameters", () => new APIParameter());
-    definitions = object.decodeObjectMap("definitions", () => new APISchemaObject());
-    securityDefinitions = object.decodeObjectMap("securityDefinitions", () => new APISecurityScheme());
+    info = object.decodeObject("info", () => APIInfo());
+    tags = object.decodeObjects("tags", () => APITag());
+    paths = object.decodeObjectMap("paths", () => APIPath());
+    responses = object.decodeObjectMap("responses", () => APIResponse());
+    parameters = object.decodeObjectMap("parameters", () => APIParameter());
+    definitions =
+        object.decodeObjectMap("definitions", () => APISchemaObject());
+    securityDefinitions = object.decodeObjectMap(
+        "securityDefinitions", () => APISecurityScheme());
   }
 
   void encode(KeyedArchive object) {