+++ lib/src/v2/header.dart
@@ -1,19 +1,19 @@
-import 'package:codable/codable.dart';
-import 'package:open_api/src/v2/property.dart';
-import 'package:open_api/src/v2/types.dart';
+import 'package:codable_forked/codable.dart';
+import 'package:open_api_forked/src/v2/property.dart';
+import 'package:open_api_forked/src/v2/types.dart';
 
 /// Represents a header in the OpenAPI specification.
 class APIHeader extends APIProperty {
   APIHeader();
 
-  String description;
-  APIProperty items;
+  String? description;
+  APIProperty? items;
 
   void decode(KeyedArchive json) {
     super.decode(json);
     description = json.decode("description");
     if (type == APIType.array) {
-      items = json.decodeObject("items", () => new APIProperty());
+      items = json.decodeObject("items", () => APIProperty());
     }
   }