+++ lib/src/v2/response.dart
@@ -1,21 +1,21 @@
-import 'package:open_api/src/object.dart';
-import 'package:open_api/src/v2/header.dart';
-import 'package:open_api/src/v2/schema.dart';
+import 'package:open_api_forked/src/object.dart';
+import 'package:open_api_forked/src/v2/header.dart';
+import 'package:open_api_forked/src/v2/schema.dart';
 
 /// Represents an HTTP response in the OpenAPI specification.
 class APIResponse extends APIObject {
   APIResponse();
 
-  String description = "";
-  APISchemaObject schema;
-  Map<String, APIHeader> headers = {};
+  String? description = "";
+  APISchemaObject? schema;
+  Map<String, APIHeader?>? headers = {};
 
   void decode(KeyedArchive object) {
     super.decode(object);
 
     description = object.decode("description");
-    schema = object.decodeObject("schema", () => new APISchemaObject());
-    headers = object.decodeObjectMap("headers", () => new APIHeader());
+    schema = object.decodeObject("schema", () => APISchemaObject());
+    headers = object.decodeObjectMap("headers", () => APIHeader());
   }
 
   void encode(KeyedArchive object) {