+++ lib/src/v2/schema.dart
@@ -1,25 +1,25 @@
-import 'package:codable/cast.dart' as cast;
-import 'package:codable/codable.dart';
-import 'package:open_api/src/v2/property.dart';
+import 'package:codable_forked/cast.dart' as cast;
+import 'package:codable_forked/codable.dart';
+import 'package:open_api_forked/src/v2/property.dart';
 
 /// Represents a schema object in the OpenAPI specification.
 class APISchemaObject extends APIProperty {
   APISchemaObject();
 
-  String title;
-  String description;
-  String example;
-  List<String> required = [];
+  String? title;
+  String? description;
+  String? example;
+  List<String?>? required = [];
   bool readOnly = false;
 
   /// Valid when type == array
-  APISchemaObject items;
+  APISchemaObject? items;
 
   /// Valid when type == null
-  Map<String, APISchemaObject> properties;
+  Map<String, APISchemaObject?>? properties;
 
   /// Valid when type == object
-  APISchemaObject additionalProperties;
+  APISchemaObject? additionalProperties;
 
   @override
   APISchemaRepresentation get representation {
@@ -42,10 +42,10 @@ class APISchemaObject extends APIProperty {
     example = json.decode("example");
     readOnly = json.decode("readOnly") ?? false;
 
-    items = json.decodeObject("items", () => new APISchemaObject());
-    additionalProperties = json.decodeObject("additionalProperties", () => new APISchemaObject());
-    properties =
-        json.decodeObjectMap("properties", () => new APISchemaObject());
+    items = json.decodeObject("items", () => APISchemaObject());
+    additionalProperties =
+        json.decodeObject("additionalProperties", () => APISchemaObject());
+    properties = json.decodeObjectMap("properties", () => APISchemaObject());
   }
 
   void encode(KeyedArchive json) {