+++ lib/src/v3/header.dart
@@ -1,6 +1,6 @@
-import 'package:codable/codable.dart';
-import 'package:open_api/src/v3/parameter.dart';
-import 'package:open_api/src/v3/schema.dart';
+import 'package:codable_forked/codable.dart';
+import 'package:open_api_forked/src/v3/parameter.dart';
+import 'package:open_api_forked/src/v3/schema.dart';
 
 /// [APIHeader] follows the structure of the [APIParameter] with the following changes:
 ///
@@ -8,7 +8,7 @@ import 'package:open_api/src/v3/schema.dart';
 /// in MUST NOT be specified, it is implicitly in header.
 /// All traits that are affected by the location MUST be applicable to a location of header (for example, style).
 class APIHeader extends APIParameter {
-  APIHeader({APISchemaObject schema}) : super.header(null, schema: schema);
+  APIHeader({APISchemaObject? schema}) : super.header(null, schema: schema);
   APIHeader.empty() : super.header(null);
 
   @override