+++ lib/src/v3/parameter.dart
@@ -1,7 +1,7 @@
-import 'package:open_api/src/object.dart';
-import 'package:open_api/src/v3/document.dart';
-import 'package:open_api/src/v3/media_type.dart';
-import 'package:open_api/src/v3/schema.dart';
+import 'package:open_api_forked/src/object.dart';
+import 'package:open_api_forked/src/v3/document.dart';
+import 'package:open_api_forked/src/v3/media_type.dart';
+import 'package:open_api_forked/src/v3/schema.dart';
 
 /// There are four possible parameter locations specified by the in field.
 ///
@@ -30,7 +30,7 @@ enum APIParameterLocation {
 }
 
 class APIParameterLocationCodec {
-  static APIParameterLocation decode(String location) {
+  static APIParameterLocation? decode(String? location) {
     switch (location) {
       case "query":
         return APIParameterLocation.query;
@@ -40,12 +40,12 @@ class APIParameterLocationCodec {
         return APIParameterLocation.path;
       case "cookie":
         return APIParameterLocation.cookie;
+      default:
+        return null;
     }
-
-    return null;
   }
 
-  static String encode(APIParameterLocation location) {
+  static String? encode(APIParameterLocation? location) {
     switch (location) {
       case APIParameterLocation.query:
         return "query";
@@ -55,8 +55,9 @@ class APIParameterLocationCodec {
         return "path";
       case APIParameterLocation.cookie:
         return "cookie";
+      default:
+        return null;
     }
-    return null;
   }
 }
 
@@ -71,12 +72,12 @@ class APIParameter extends APIObject {
       this.schema,
       this.content,
       this.style,
-      bool required,
-      bool deprecated,
-      bool allowEmptyValue,
-      bool explode,
-      bool allowReserved}) {
-    this.isRequired = required;
+      bool? isRequired,
+      bool? deprecated,
+      bool? allowEmptyValue,
+      bool? explode,
+      bool? allowReserved}) {
+    this.isRequired = isRequired;
     this.isDeprecated = deprecated;
     this.allowEmptyValue = allowEmptyValue;
     this.allowReserved = allowReserved;
@@ -88,12 +89,12 @@ class APIParameter extends APIObject {
       this.schema,
       this.content,
       this.style,
-      bool required,
-      bool deprecated,
-      bool allowEmptyValue,
-      bool explode,
-      bool allowReserved}) {
-    this.isRequired = required;
+      bool? isRequired,
+      bool? deprecated,
+      bool? allowEmptyValue,
+      bool? explode,
+      bool? allowReserved}) {
+    this.isRequired = isRequired;
     this.isDeprecated = deprecated;
     this.allowEmptyValue = allowEmptyValue;
     this.allowReserved = allowReserved;
@@ -106,12 +107,12 @@ class APIParameter extends APIObject {
       this.schema,
       this.content,
       this.style,
-      bool required,
-      bool deprecated,
-      bool allowEmptyValue,
-      bool explode,
-      bool allowReserved}) {
-    this.isRequired = required;
+      bool? isRequired,
+      bool? deprecated,
+      bool? allowEmptyValue,
+      bool? explode,
+      bool? allowReserved}) {
+    this.isRequired = isRequired;
     this.isDeprecated = deprecated;
     this.allowEmptyValue = allowEmptyValue;
     this.allowReserved = allowReserved;
@@ -121,7 +122,7 @@ class APIParameter extends APIObject {
 
   APIParameter.path(this.name)
       : location = APIParameterLocation.path,
-        schema = new APISchemaObject.string(),
+        schema = APISchemaObject.string(),
         _required = true;
 
   APIParameter.cookie(this.name,
@@ -129,12 +130,12 @@ class APIParameter extends APIObject {
       this.schema,
       this.content,
       this.style,
-      bool required,
-      bool deprecated,
-      bool allowEmptyValue,
-      bool explode,
-      bool allowReserved}) {
-    this.isRequired = required;
+      bool? isRequired,
+      bool? deprecated,
+      bool? allowEmptyValue,
+      bool? explode,
+      bool? allowReserved}) {
+    this.isRequired = isRequired;
     this.isDeprecated = deprecated;
     this.allowEmptyValue = allowEmptyValue;
     this.allowReserved = allowReserved;
@@ -148,83 +149,85 @@ class APIParameter extends APIObject {
   /// If in is "path", the name field MUST correspond to the associated path segment from the path field in [APIDocument.paths]. See Path Templating for further information.
   /// If in is "header" and the name field is "Accept", "Content-Type" or "Authorization", the parameter definition SHALL be ignored.
   /// For all other cases, the name corresponds to the parameter name used by the in property.
-  String name;
+  String? name;
 
   /// A brief description of the parameter.
   ///
   /// This could contain examples of use. CommonMark syntax MAY be used for rich text representation.
-  String description;
+  String? description;
 
   /// Determines whether this parameter is mandatory.
   ///
   /// If the parameter location is "path", this property is REQUIRED and its value MUST be true. Otherwise, the property MAY be included and its default value is false.
-  bool get isRequired => (location == APIParameterLocation.path ? true : (_required ?? false));
+  bool get isRequired =>
+      (location == APIParameterLocation.path ? true : _required ?? false);
 
-  set isRequired(bool f) {
+  set isRequired(bool? f) {
     _required = f;
   }
 
-  bool _required;
+  bool? _required = false;
 
   /// Specifies that a parameter is deprecated and SHOULD be transitioned out of usage.
   bool get isDeprecated => _deprecated ?? false;
 
-  set isDeprecated(bool f) {
+  set isDeprecated(bool? f) {
     _deprecated = f;
   }
 
-  bool _deprecated;
+  bool? _deprecated = false;
 
   /// The location of the parameter.
   ///
   /// REQUIRED. Possible values are "query", "header", "path" or "cookie".
-  APIParameterLocation location;
+  /// EXCEPT when used as response header 🤦️
+  APIParameterLocation? location;
 
   /// The schema defining the type used for the parameter.
-  APISchemaObject schema;
+  APISchemaObject? schema;
 
   // Sets the ability to pass empty-valued parameters.
   //
   // This is valid only for query parameters and allows sending a parameter with an empty value. Default value is false. If style is used, and if behavior is n/a (cannot be serialized), the value of allowEmptyValue SHALL be ignored.
-  bool get allowEmptyValue => _allowEmptyValue ?? false;
+  bool? get allowEmptyValue => _allowEmptyValue;
 
-  set allowEmptyValue(bool f) {
+  set allowEmptyValue(bool? f) {
     _allowEmptyValue = f;
   }
 
-  bool _allowEmptyValue;
+  bool? _allowEmptyValue = false;
 
   /// Describes how the parameter value will be serialized depending on the type of the parameter value.
   ///
   /// Default values (based on value of in): for query - form; for path - simple; for header - simple; for cookie - form.
-  String style;
+  String? style;
 
   /// When this is true, parameter values of type array or object generate separate parameters for each value of the array or key-value pair of the map.
   ///
   /// For other types of parameters this property has no effect. When style is form, the default value is true. For all other styles, the default value is false.
-  bool get explode => _explode ?? false;
+  bool? get explode => _explode;
 
-  set explode(bool f) {
+  set explode(bool? f) {
     _explode = f;
   }
 
-  bool _explode;
+  bool? _explode = false;
 
   /// Determines whether the parameter value SHOULD allow reserved characters, as defined by RFC3986 :/?#[]@!$&'()*+,;= to be included without percent-encoding.
   ///
   /// This property only applies to parameters with an in value of query. The default value is false.
-  bool get allowReserved => _allowReserved ?? false;
+  bool? get allowReserved => _allowReserved;
 
-  set allowReserved(bool f) {
+  set allowReserved(bool? f) {
     _allowReserved = f;
   }
 
-  bool _allowReserved;
+  bool? _allowReserved = false;
 
   /// A map containing the representations for the parameter.
   ///
   /// The key is the media type and the value describes it. The map MUST only contain one entry.
-  Map<String, APIMediaType> content;
+  Map<String, APIMediaType?>? content;
 
   // Currently missing:
   // example, examples
@@ -240,18 +243,19 @@ class APIParameter extends APIObject {
     _deprecated = object.decode("deprecated");
     _allowEmptyValue = object.decode("allowEmptyValue");
 
-    schema = object.decodeObject("schema", () => new APISchemaObject());
+    schema = object.decodeObject("schema", () => APISchemaObject());
     style = object.decode("style");
     _explode = object.decode("explode");
     _allowReserved = object.decode("allowReserved");
-    content = object.decodeObjectMap("content", () => new APIMediaType());
+    content = object.decodeObjectMap("content", () => APIMediaType());
   }
 
   void encode(KeyedArchive object) {
     super.encode(object);
 
     if (name == null || location == null) {
-      throw new ArgumentError("APIParameter must have non-null values for: 'name', 'location'.");
+      throw ArgumentError(
+          "APIParameter must have non-null values for: 'name', 'location'.");
     }
 
     object.encode("name", name);